+++ decidim-admin/app/controllers/concerns/decidim/admin/participatory_space_export.rb
@@ -1,0 +1,43 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Admin
+    module ParticipatorySpaceExport
+      extend ActiveSupport::Concern
+
+      included do
+        helper_method :exportable_space
+
+        def create
+          enforce_permission_to :create, :export_space, participatory_space: exportable_space
+
+          ExportParticipatorySpaceJob.perform_later(current_user, exportable_space, "participatory_processes", default_format)
+
+          flash[:notice] = t("decidim.admin.exports.notice")
+
+          redirect_back(fallback_location: after_export_path)
+        end
+
+        # Public: The only method to be implemented at the controller. You need to
+        # return the space that will be exported.
+        def exportable_space
+          raise NotImplementedError
+        end
+
+        # Public: Returns a String or Object that will be passed to `redirect_to` after
+        # exporing a space. By default it redirects to the root_path.
+        #
+        # It can be redefined at controller level if you need to redirect elsewhere.
+        def after_export_path
+          decidim.root_path
+        end
+
+        private
+
+        def default_format
+          "JSON"
+        end
+      end
+    end
+  end
+end