+++ decidim-participatory_processes/lib/decidim/participatory_processes/participatory_space.rb
@@ -33,6 +33,14 @@ Decidim.register_participatory_space(:participatory_processes) do |participatory
     context.layout = "layouts/decidim/admin/participatory_process"
   end
 
+  participatory_space.exports :participatory_processes do |export|
+    export.collection do |participatory_process|
+      Decidim::ParticipatoryProcess.where(id: participatory_process.id)
+    end
+
+    export.serializer Decidim::ParticipatoryProcesses::ParticipatoryProcessSerializer
+  end
+
   participatory_space.seeds do
     organization = Decidim::Organization.first
     seeds_root = File.join(__dir__, "..", "..", "..", "db", "seeds")